Split the code into three parts, split the screenshots making two more, basically rephrased the page completely
