+++ docpages/example_programs/the_basics/editing-channels-and-messages.md
@@ -1,24 +1,42 @@
 \page editing-channels-and-messages Editing Channels and Messages
 
-Sometimes we need to update an object, such as a message (whether it's plain text or an embed) or a channel. At first, it might seem confusing, but it's actually really simple! You need an object with all of the properties being identical to the existing one. Say you're editing a message. You need to have an object with its ID the same as the one in Discord. Then you replace what you need, such as its content.
+Sometimes we need to update an object, such as a message (whether it's plain text or an embed) or a channel. At first, it might seem confusing, but it's actually really simple! You need an object with all the properties being identical to the existing one. Say you're editing a message. You need to have an object with its ID the same as the one in Discord. Then you replace what you need, such as its content.
 
 \note Your bot can't edit messages sent by others!
 \note This example uses callback functions and embeds. To see more information about them, visit \ref callback-functions and \ref embed-message.
 
-\include{cpp} editing_messages.cpp
+Here we send a message and edit it after. To do so, we first reply to the command `msg-send` with some text, "This is a message" in our case. As described above, on the next step the message object is taken and the text is replaced with whatever the user desires.
 
-Before editing:
+\include{cpp} editing_messages1.cpp
+
+Before editing the message:
 
 \image html stuff_edit1.png
 
-After editing:
+After editing the message:
 
 \image html stuff_edit2.png
 
-Now let's send the embed:
+Now we'll want to edit an existing channel - its name in this case. This works similarly to how messages are edited.
+
+\include{cpp} editing_messages2.cpp
+
+Before editing the channel:
 
 \image html stuff_edit3.png
 
-And finally, edit it:
+After editing the channel:
 
 \image html stuff_edit4.png
+
+Now let's send the embed and edit it. It's close but a bit more different than before. Say if a message has one `content` field, it can have a few `embed` fields, up to 10 to be precise. So we first get the embed we want and edit and change its description.
+
+\include{cpp} editing_messages3.cpp
+
+Before editing the embed:
+
+\image html stuff_edit5.png
+
+Finally, after editing the embed:
+
+\image html stuff_edit6.png
